SmthToDo is my first project based on React Native for iOS and Android.
To run the application please use expo.

Expo is a platform and framework for developing universal React based applications. Expo provides a set of services on React Native to progress and uses native platforms, which help the developer to implement and build the application on iOS, Android and web. The code used for the three platforms is the same (Expo documentation).